requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
Pillow==7.1.0
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
